Lab-01-Tech-Insights/mcp-scripts/requirements.txt
httpx
feedparser
beautifulsoup4
PyYAML
python-dotenv
google-genai
